requirement.txt
fastapi
fastapi_utils
uvicorn
aiofiles
mutagen
numpy
pydantic
PyJWT
pytz
rapidfuzz
requests
scipy
sqlalchemy
brotlipy
requests-async
shapely
watchdog
pydub
boto3
asyncpg
python-multipart
passlib[bcrypt]
python-jose
psycopg2-binary
